chore: refactor install java out of extension.ts
