+++ src/consts.ts
@@ -1,0 +1,4 @@
+export const openSettingsAction = "Open settings";
+
+export const installJava11Action = "Install Java (JDK 11)";
+export const installJava17Action = "Install Java (JDK 17)";
